+++ ga_prdxpr/src/ga_prdxpr_main.cpp
@@ -5,6 +5,8 @@
 
 #include "ga_prd_ega2dp.hpp"
 
+#include "ga_prdxpr_transformer.hpp" // just for ParseError
+
 
 // multivector basis, product rules, simplification rules and coefficients need to be
 // provided in header files included above by the user
@@ -13,6 +15,9 @@ int main()
 {
 
     try {
+
+        // regular products with two operands
+
         // ega2d
         generate_and_print_ega2d_gpr();    // geometric product
         generate_and_print_ega2d_wdg();    // wegde product
@@ -42,8 +47,15 @@ int main()
 
         generate_and_print_ega2dp_rdot(); // regressive scalar product
         generate_and_print_ega2dp_rwdg(); // regressive wedge product
+
+        // sandwich products of type rotor * object * rev(rotor)
+        generate_and_print_ega2d_rotor();
+        generate_and_print_ega3d_rotor();
     }
 
+    catch (ParseError const& e) {
+        fmt::println("Parse error: {}", e.what());
+    }
     catch (std::exception const& e) {
         fmt::println("Exception: {}", e.what());
     }